backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pysqlite3-binary==0.5.4
python-dotenv==1.0.1
sqlparse==0.5.3
